+++ tests/test_exceptions.py
@@ -1,9 +1,10 @@
-from wannierberri import __utility as util
+from wannierberri import utility as util
 import pytest
 import numpy as np
 import wannierberri as wberri
 from .common_systems import create_files_tb
 from wannierberri import calculators as calc
+from wannierberri.fourier.fft import FFT_R_to_k, execute_fft
 
 
 def test_utility_str2bool():
@@ -20,10 +21,10 @@ def test_utility_FFT():
     inp = np.random.random((5, 4, 3, 5)) * (1 + 2.j)
     axes = (0, 2)
     for fft in "fftw", "numpy", "FfTw", "nUmPY":
-        util.execute_fft(inp, axes, fftlib=fft)
+        execute_fft(inp, axes, fftlib=fft)
     for fft in "fftlib", "np", "idkwhat", "dummy":
         with pytest.raises(ValueError, match=f"unknown type of fftlib : {fft}"):
-            util.execute_fft(inp, axes, fftlib=fft)
+            execute_fft(inp, axes, fftlib=fft)
 
 
 def test_utility_FFT_R_to_k():
@@ -31,18 +32,18 @@ def test_utility_FFT_R_to_k():
     iRvec = np.random.randint(100, size=(20, 3))
     num_wann = 5
     for lib in 'fftw', 'numpy', 'NuMpY', 'fftW':
-        util.FFT_R_to_k(iRvec, NKFFT, num_wann, fftlib=lib)
+        FFT_R_to_k(iRvec, NKFFT, num_wann, fftlib=lib)
 
     for lib in 'unknonw', 'fftlib', 'nump', 'np', "NP":
         with pytest.raises(AssertionError, match=f"fftlib '{lib.lower()}' is unknown/not supported"):
-            util.FFT_R_to_k(iRvec, NKFFT, num_wann, fftlib=lib)
+            FFT_R_to_k(iRvec, NKFFT, num_wann, fftlib=lib)
 
     shape = (num_wann, num_wann) + NKFFT + (3, 3, 3)
     AAA_K = np.random.random(shape) + 1j * np.random.random(shape)
 
-    util.FFT_R_to_k(iRvec, NKFFT, num_wann, fftlib="numpy").transform(AAA_K)
+    FFT_R_to_k(iRvec, NKFFT, num_wann, fftlib="numpy").transform(AAA_K)
     with pytest.raises(RuntimeError, match="FFT.transform should not be called for slow FT"):
-        util.FFT_R_to_k(iRvec, NKFFT, num_wann, fftlib="slow").transform(AAA_K)
+        FFT_R_to_k(iRvec, NKFFT, num_wann, fftlib="slow").transform(AAA_K)
 
 
 @pytest.mark.parametrize("ibands", [[5, 6], [4, 6, 7, 8]])
