+++ wannierberri/system/system_tb.py
@@ -15,7 +15,7 @@ import numpy as np
 import os
 from termcolor import cprint
 
-from .rvectors import Rvectors
+from ..fourier.rvectors import Rvectors
 from .system_R import System_R
